workspot-bengaluru/backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
djangorestframework-simplejwt>=5.3.0
Pillow>=10.0.0
django-filter>=23.5
gunicorn>=21.2.0
